applications: connectivity_bridge: Remove misused PM_POLICY_APP option
This application does not implement functions pm_policy_next_state()
pm_policy_low_power_devices() what is required when the PM_POLICY_APP
option is used. This option, in an outdated form with SYS_ prefix,
is only used there in a check if power states should be changed for
UART. Remove this option from this application configuration then
and correct the check.

Signed-off-by: Andrzej Głąbek <andrzej.glabek@nordicsemi.no>